refactor: update actions and reducer
