+++ src/components/Card/HashCard/index.tsx
@@ -6,7 +6,7 @@ import i18n from '../../../utils/i18n'
 import { isMobile } from '../../../utils/screen'
 import { adaptPCEllipsis, adaptMobileEllipsis } from '../../../utils/string'
 import { copyElementValue } from '../../../utils/util'
-import { AppActions } from '../../../contexts/providers/reducer'
+import { AppActions } from '../../../contexts/actions'
 import SmallLoading from '../../Loading/SmallLoading'
 import { useDispatch } from '../../../contexts/providers'
 import { HashCardPanel, LoadingPanel } from './styled'
