+++ src/components/Dropdown/Language/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import i18n, { currentLanguage, changeLanguage } from '../../../utils/i18n'
 import { useDispatch } from '../../../contexts/providers'
-import { AppActions } from '../../../contexts/providers/reducer'
+import { AppActions } from '../../../contexts/actions'
 import { LanguagePanel } from './styled'
 import SimpleButton from '../../SimpleButton'
