+++ src/components/Header/index.tsx
@@ -6,7 +6,7 @@ import MobileMenuCloseIcon from '../../assets/menu_close_mobile.png'
 import { HeaderPanel, HeaderEmptyPanel, HeaderMobileMenuPanel, HeaderLogoPanel } from './styled'
 import { isMobile, isScreen750to1440 } from '../../utils/screen'
 import { useAppState, useDispatch } from '../../contexts/providers/index'
-import { ComponentActions } from '../../contexts/providers/reducer'
+import { ComponentActions } from '../../contexts/actions'
 import MenusComp from './MenusComp'
 import { SearchComp } from './SearchComp'
 import LanguageComp from './LanguageComp'
