+++ src/components/Search/DaoSearch/index.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next'
 import GreenSearchLogo from '../../../assets/search_green.png'
 import BlueSearchLogo from '../../../assets/search_blue.png'
 import i18n from '../../../utils/i18n'
-import { AppActions } from '../../../contexts/providers/reducer'
+import { AppActions } from '../../../contexts/actions'
 import { isMainnet } from '../../../utils/chain'
 import { searchNervosDaoTransactions, getNervosDaoTransactions } from '../../../service/app/nervosDao'
 import { useDispatch } from '../../../contexts/providers'
