+++ src/components/Search/UDTSearch/index.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next'
 import SearchLogo from '../../../assets/search_black.png'
 import ClearLogo from '../../../assets/clear.png'
 import i18n from '../../../utils/i18n'
-import { AppActions, ComponentActions } from '../../../contexts/providers/reducer'
+import { AppActions, ComponentActions } from '../../../contexts/actions'
 import { useDispatch, useAppState } from '../../../contexts/providers'
 import { getSimpleUDTTransactionsWithAddress } from '../../../service/app/udt'
 import { isMobile } from '../../../utils/screen'
