+++ src/components/Search/index.tsx
@@ -11,7 +11,8 @@ import ClearLogo from '../../assets/clear.png'
 import { addPrefixForHash } from '../../utils/string'
 import i18n from '../../utils/i18n'
 import { HttpErrorCode, SearchFailType } from '../../utils/const'
-import { AppDispatch, AppActions, ComponentActions } from '../../contexts/providers/reducer'
+import { AppDispatch } from '../../contexts/reducer'
+import { AppActions, ComponentActions } from '../../contexts/actions'
 import { isMobile } from '../../utils/screen'
 import { useAppState, useDispatch } from '../../contexts/providers'
 import { isMainnet } from '../../utils/chain'
