+++ src/components/Text/CopyTooltipText/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import i18n from '../../../utils/i18n'
 import { copyElementValue } from '../../../utils/util'
-import { AppActions } from '../../../contexts/providers/reducer'
+import { AppActions } from '../../../contexts/actions'
 import { useDispatch } from '../../../contexts/providers'
 import SimpleButton from '../../SimpleButton'
