+++ src/contexts/reducer/index.ts
@@ -1,78 +1,4 @@
-export enum AppActions {
-  ResizeWindow = 'resizeWindow',
-  UpdateLoading = 'updateLoading',
-  UpdateSecondLoading = 'updateSecondLoading',
-  UpdateModal = 'updateModal',
-  ShowToastMessage = 'showToastMessage',
-  UpdateAppErrors = 'updateAppErrors',
-  UpdateNodeVersion = 'updateNodeVersion',
-  UpdateTipBlockNumber = 'updateTipBlockNumber',
-  UpdateAppLanguage = 'updateAppLanguage',
-}
-
-export enum PageActions {
-  UpdateAddress = 'updateAddress',
-  UpdateAddressTransactions = 'updateAddressTransactions',
-  UpdateAddressTotal = 'updateAddressTotal',
-  UpdateAddressStatus = 'updateAddressStatus',
-  UpdateAddressTransactionsStatus = 'updateAddressTransactionsStatus',
-
-  UpdateHomeBlocks = 'updateHomeBlocks',
-
-  UpdateBlockList = 'updateBlockList',
-  UpdateBlockListTotal = 'updateBlockListTotal',
-
-  UpdateBlock = 'updateBlock',
-  UpdateBlockTransactions = 'updateBlockTransactions',
-  UpdateBlockTotal = 'updateBlockTotal',
-  UpdateBlockStatus = 'updateBlockStatus',
-
-  UpdateTransaction = 'updateTransaction',
-  UpdateTransactionStatus = 'updateTransactionStatus',
-  UpdateTransactionScriptFetched = 'updateTransactionScriptFetched',
-  UpdateTransactions = 'updateTransactions',
-  UpdateTransactionsTotal = 'updateTransactionsTotal',
-
-  UpdateStatistics = 'updateStatistics',
-
-  UpdateStatisticDifficultyHashRate = 'updateStatisticDifficultyHashRate',
-  UpdateStatisticDifficultyUncleRate = 'updateStatisticDifficultyUncleRate',
-  UpdateStatisticDifficulty = 'updateStatisticDifficulty',
-  UpdateStatisticHashRate = 'updateStatisticHashRate',
-  UpdateStatisticUncleRate = 'updateStatisticUncleRate',
-  UpdateStatisticTransactionCount = 'updateStatisticTransactionCount',
-  UpdateStatisticAddressCount = 'updateStatisticAddressCount',
-  UpdateStatisticTotalDaoDeposit = 'updateStatisticTotalDaoDeposit',
-  UpdateStatisticNewDaoDeposit = 'updateStatisticNewDaoDeposit',
-  UpdateStatisticNewDaoWithdraw = 'updateStatisticNewDaoWithdraw',
-  UpdateStatisticCirculationRatio = 'updateStatisticCirculationRatio',
-  UpdateStatisticCellCount = 'updateStatisticCellCount',
-  UpdateStatisticAddressBalanceRank = 'updateStatisticAddressBalanceRank',
-  UpdateStatisticBalanceDistribution = 'updateStatisticBalanceDistribution',
-  UpdateStatisticTxFeeHistory = 'updateStatisticTxFeeHistory',
-  UpdateStatisticBlockTimeDistribution = 'updateStatisticBlockTimeDistribution',
-  UpdateStatisticOccupiedCapacity = 'updateStatisticOccupiedCapacity',
-  UpdateStatisticEpochTimeDistribution = 'updateStatisticEpochTimeDistribution',
-  UpdateStatisticEpochLengthDistribution = 'updateStatisticEpochLengthDistribution',
-
-  UpdateNervosDao = 'updateNervosDao',
-  UpdateNervosDaoTransactions = 'updateNervosDaoTransactions',
-  UpdateNervosDaoTransactionsTotal = 'updateNervosDaoTransactionsTotal',
-  UpdateNervosDaoDepositors = 'updateNervosDaoDepositors',
-  UpdateNervosDaoStatus = 'updateNervosDaoStatus',
-
-  UpdateUDT = 'updateUDT',
-  UpdateUDTTransactions = 'updateUDTTransactions',
-  UpdateUDTTransactionsTotal = 'updateUDTTransactionsTotal',
-  UpdateUDTStatus = 'updateUDTStatus',
-}
-
-export enum ComponentActions {
-  UpdateHeaderSearchEditable = 'updateHeaderSearchEditable',
-  UpdateHeaderMobileMenuVisible = 'updateHeaderMobileMenuVisible',
-}
-
-export type StateActions = AppActions | PageActions | ComponentActions
+import StateActions, { AppActions, PageActions, ComponentActions } from '../actions'
 
 export type AppDispatch = React.Dispatch<{ type: StateActions; payload: any }> // TODO: add type of payload
 export type StateWithDispatch = State.AppState & { dispatch: AppDispatch }
