+++ src/pages/BlockDetail/index.tsx
@@ -5,7 +5,7 @@ import BlockHashCard from '../../components/Card/HashCard'
 import Content from '../../components/Content'
 import Error from '../../components/Error'
 import { useDispatch, useAppState } from '../../contexts/providers'
-import { PageActions, AppActions } from '../../contexts/providers/reducer'
+import { PageActions, AppActions } from '../../contexts/actions'
 import { getBlock } from '../../service/app/block'
 import { PageParams, LOADING_WAITING_TIME } from '../../utils/const'
 import i18n from '../../utils/i18n'
