+++ src/pages/SimpleUDT/index.tsx
@@ -6,7 +6,7 @@ import SimpleUDTHashCard from '../../components/Card/HashCard'
 import Error from '../../components/Error'
 import Content from '../../components/Content'
 import { useAppState, useDispatch } from '../../contexts/providers/index'
-import { PageActions, AppActions } from '../../contexts/providers/reducer'
+import { PageActions, AppActions } from '../../contexts/actions'
 import { getTipBlockNumber } from '../../service/app/address'
 import { PageParams, LOADING_WAITING_TIME } from '../../utils/const'
 import i18n from '../../utils/i18n'
