+++ src/pages/StatisticsChart/activities/AddressBalanceRank.tsx
@@ -1,7 +1,8 @@
 import React, { useEffect, useCallback } from 'react'
 import 'default-passive-events'
 import { getStatisticAddressBalanceRank } from '../../../service/app/charts/activities'
-import { PageActions, AppDispatch } from '../../../contexts/providers/reducer'
+import { AppDispatch } from '../../../contexts/reducer'
+import { PageActions } from '../../../contexts/actions'
 import { useAppState, useDispatch } from '../../../contexts/providers'
 import i18n, { currentLanguage } from '../../../utils/i18n'
 import { handleAxis } from '../../../utils/chart'
