+++ src/pages/StatisticsChart/activities/TxFeeHistory.tsx
@@ -8,7 +8,8 @@ import { parseDateNoTime } from '../../../utils/date'
 import { isMobile } from '../../../utils/screen'
 import { ChartColors } from '../../../utils/const'
 import { ChartLoading, ReactChartCore, ChartPage } from '../common/ChartComp'
-import { PageActions, AppDispatch } from '../../../contexts/providers/reducer'
+import { AppDispatch } from '../../../contexts/reducer'
+import { PageActions } from '../../../contexts/actions'
 import { shannonToCkbDecimal } from '../../../utils/util'
 
 const gridThumbnail = {
