+++ src/pages/StatisticsChart/block/EpochLengthDistribution.tsx
@@ -4,7 +4,8 @@ import i18n, { currentLanguage } from '../../../utils/i18n'
 import { isMobile } from '../../../utils/screen'
 import { ChartColors } from '../../../utils/const'
 import { ChartLoading, ReactChartCore, ChartPage } from '../common/ChartComp'
-import { PageActions, AppDispatch } from '../../../contexts/providers/reducer'
+import { AppDispatch } from '../../../contexts/reducer'
+import { PageActions } from '../../../contexts/actions'
 import { getStatisticEpochLengthDistribution } from '../../../service/app/charts/block'
 import { localeNumberString } from '../../../utils/number'
 import { handleStepGroupAxis } from '../../../utils/chart'
