+++ src/pages/StatisticsChart/mining/DifficultyUncleRate.tsx
@@ -8,7 +8,8 @@ import { handleDifficulty } from '../../../utils/number'
 import { isMobile } from '../../../utils/screen'
 import { ChartColors } from '../../../utils/const'
 import { ChartLoading, ReactChartCore, ChartPage } from '../common/ChartComp'
-import { PageActions, AppDispatch } from '../../../contexts/providers/reducer'
+import { AppDispatch } from '../../../contexts/reducer'
+import { PageActions } from '../../../contexts/actions'
 
 const gridThumbnail = {
   left: '4%',
