+++ src/pages/Transaction/TransactionCellScript/index.tsx
@@ -15,7 +15,8 @@ import {
 import CopyIcon from '../../../assets/copy_green.png'
 import CopyBlueIcon from '../../../assets/copy_blue.png'
 import i18n from '../../../utils/i18n'
-import { AppActions, AppDispatch, PageActions } from '../../../contexts/providers/reducer'
+import { AppDispatch } from '../../../contexts/reducer'
+import { AppActions, PageActions } from '../../../contexts/actions'
 import SmallLoading from '../../../components/Loading/SmallLoading'
 import { isMainnet } from '../../../utils/chain'
 import { useDispatch, useAppState } from '../../../contexts/providers'
