+++ src/service/app/charts/mining.ts
@@ -6,7 +6,8 @@ import {
   fetchStatisticHashRate,
   fetchStatisticUncleRate,
 } from '../../http/fetcher'
-import { AppDispatch, PageActions } from '../../../contexts/providers/reducer'
+import { AppDispatch } from '../../../contexts/reducer'
+import { PageActions } from '../../../contexts/actions'
 
 export const getStatisticDifficultyHashRate = (dispatch: AppDispatch) => {
   fetchStatisticDifficultyHashRate().then(
