+++ src/service/app/nodeInfo.ts
@@ -1,5 +1,6 @@
 import { fetchNodeVersion } from '../http/fetcher'
-import { AppActions, AppDispatch } from '../../contexts/providers/reducer'
+import { AppDispatch } from '../../contexts/reducer'
+import { AppActions } from '../../contexts/actions'
 
 export const initNodeVersion = (dispatch: AppDispatch) => {
   fetchNodeVersion().then((wrapper: Response.Wrapper<State.NodeVersion> | null) => {
