+++ src/service/app/statistics.ts
@@ -1,5 +1,6 @@
 import { fetchStatistics } from '../http/fetcher'
-import { AppDispatch, PageActions } from '../../contexts/providers/reducer'
+import { AppDispatch } from '../../contexts/reducer'
+import { PageActions } from '../../contexts/actions'
 
 export const getStatistics = (dispatch: AppDispatch) => {
   fetchStatistics().then((wrapper: Response.Wrapper<State.Statistics> | null) => {
